+++ client-app-android/src/uimock/java/de/njsm/stocks/client/business/UimockBusinessModule.java
@@ -221,4 +221,7 @@ public interface UimockBusinessModule {
 
     @Binds
     EanNumberAssignmentInteractor EanNumberAssignmentInteractor(InMemoryEanNumberAssignmentInteractorImpl impl);
+
+    @Binds
+    RecipeDetailInteractor RecipeDetailInteractor(InMemoryRecipeDetailInteractorImpl impl);
 }