+++ client-database-android/src/main/java/de/njsm/stocks/client/database/RecipeDao.java
@@ -22,6 +22,7 @@ package de.njsm.stocks.client.database;
 
 import androidx.room.Dao;
 import androidx.room.Query;
+import de.njsm.stocks.client.business.entities.RecipeForDetailsBaseData;
 import de.njsm.stocks.client.business.entities.RecipeForListingBaseData;
 import io.reactivex.rxjava3.core.Observable;
 
@@ -54,4 +55,49 @@ abstract class RecipeDao {
             "join current_unit u on u.id = s.unit " +
             "order by i.id")
     abstract Observable<List<RecipeListRepositoryImpl.RecipeIngredientAmountBaseData>> getIngredientsRequiredAmount();
+
+    @Query("select i.id, food.name as foodName, u.abbreviation, s.scale, count(*) as amount " +
+            "from current_recipe_ingredient i " +
+            "join current_food food on food.id = i.ingredient " +
+            "join current_food_item f on f.of_type = i.ingredient " +
+            "join current_scaled_unit s on f.unit = s.id " +
+            "join current_unit u on u.id = s.unit " +
+            "where i.recipe = :recipeId " +
+            "group by i.recipe, i.id, u.id, s.scale " +
+            "order by i.id")
+    abstract Observable<List<RecipeDetailRepositoryImpl.RecipeFoodForDetailsBaseData>> getIngredientsPresentAmountsOf(int recipeId);
+
+    @Query("select i.id, f.name as foodName, u.abbreviation, s.scale, i.amount " +
+            "from current_recipe_ingredient i " +
+            "join current_food f on f.id = i.ingredient " +
+            "join current_scaled_unit s on i.unit = s.id " +
+            "join current_unit u on u.id = s.unit " +
+            "where i.recipe = :recipeId " +
+            "order by i.id")
+    abstract Observable<List<RecipeDetailRepositoryImpl.RecipeFoodForDetailsBaseData>> getIngredientsRequiredAmountOf(int recipeId);
+
+    @Query("select i.id, food.name as foodName, u.abbreviation, s.scale, count(*) as amount " +
+            "from current_recipe_product i " +
+            "join current_food food on food.id = i.product " +
+            "join current_food_item f on f.of_type = i.product " +
+            "join current_scaled_unit s on f.unit = s.id " +
+            "join current_unit u on u.id = s.unit " +
+            "where i.recipe = :recipeId " +
+            "group by i.recipe, i.id, u.id, s.scale " +
+            "order by i.id")
+    abstract Observable<List<RecipeDetailRepositoryImpl.RecipeFoodForDetailsBaseData>> getProductsPresentAmountsOf(int recipeId);
+
+    @Query("select i.id, f.name as foodName, u.abbreviation, s.scale, i.amount " +
+            "from current_recipe_product i " +
+            "join current_food f on f.id = i.product " +
+            "join current_scaled_unit s on i.unit = s.id " +
+            "join current_unit u on u.id = s.unit " +
+            "where i.recipe = :recipeId " +
+            "order by i.id")
+    abstract Observable<List<RecipeDetailRepositoryImpl.RecipeFoodForDetailsBaseData>> getProductsProducedAmountOf(int recipeId);
+
+    @Query("select r.id, r.name, r.instructions, r.duration " +
+            "from current_recipe r " +
+            "where r.id = :recipeId")
+    abstract Observable<RecipeForDetailsBaseData> getRecipe(int recipeId);
 }