+++ client-navigation-android/src/main/java/de/njsm/stocks/client/navigation/NavigationModule.java
@@ -128,4 +128,7 @@ public interface NavigationModule {
 
     @Binds
     FoodEanNumberAssignmentNavigator FoodEanNumberAssignmentNavigator(FoodEanNumberAssignmentNavigatorImpl impl);
+
+    @Binds
+    RecipeDetailNavigator RecipeDetailNavigator(RecipeDetailNavigatorImpl impl);
 }